# docs/Swagger.yaml
swagger: '2.0'
info:
  description: |
    Mushroom Cultivation System
  version: pre-alpha-1.0
  title: Mushroom++
paths:
...
  /user:
    post:
      summary: creates a new user
      parameters:
      - in: body
        name: User
        required: true
        description: User
        schema:
          $ref: '#/definitions/user'
      produces:
      - text/plain
      responses:
        200:
          description: Created
        409:
          description: "Conflict, descriptive error message detailing if the username is taken or invalid, or if the password is too weak or invalid."
          schema:
            example: "username \"Bob\" taken"
  /user/name/{username}:
    get:
      summary: returns the user with the given username
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: username
        in: path
        description: username of the user to return
        required: true
        type: string
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/user'
        403:
          description: Unauthorized
        404:
          description: Not Found
  /user/name:
    get:
      summary: returns the authenticated user
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/user'
        403:
          description: Unauthorized
  /user/key/{user_key}:
    delete:
      summary: deletes the user with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: user_key
        in: path
        description: key of the User
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        200:
          description: Deleted
        403:
          description: Unauthorized
        404:
          description: Not Found
  /user/key/{user_key}/username:
    patch:
      summary: changes the username of the user with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: user_key
        in: path
        description: key of the User
        required: true
        type: integer
      - name: username
        in: body
        description: "new username\nUser Model could be used as it has this field, while the other fields can be ignored."
        required: true
        schema:
          type: object
          properties:
            username:
              type: string
              example: Bob
      produces:
      - text/plain
      responses:
        200:
          description: Username Changed
        403:
          description: Unauthorized
        404:
          description: Not Found
        409:
          description: "Conflict, descriptive error message detailing if the username is taken or invalid."
          schema:
            example: "username \"Bob\" taken"
  /user/key/{user_key}/password:
    patch:
      summary: changes the password of the user with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: user_key
        in: path
        description: key of the User
        required: true
        type: integer
      - name: password
        in: body
        description: "new password\nUser Model could be used as it has this field, while the other fields can be ignored."
        required: true
        schema:
          type: object
          properties:
            password:
              type: string
              example: kolbiharcos1234
      produces:
      - text/plain
      responses:
        200:
          description: Password Changed
        403:
          description: Unauthorized
        404:
          description: Not Found
        409:
          description: "Conflict, descriptive error message detailing if the password is too weak or invalid."
          schema:
            example: "the password is too short"
...
  /specimen/key/{specimen_key}/status:
    post:
      summary: creates a new status entry associated with the specimen with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: specimen_key
        in: path
        description: key of the Specimen
        required: true
        type: integer
      - in: body
        name: Status Entry
        required: true
        description: Status Entry
        schema:
          $ref: '#/definitions/status_entry'
      produces:
      - text/plain
      responses:
        200:
          description: Created
        403:
          description: Unauthorized (can be because of the Specimen is marked as inactive, aka discarded, aka deleted)
        404:
          description: Not Found
        409:
          description: "Conflict, descriptive error message detailing if the stage_name is invalid."
          schema:
            example: "stage_name \"Fruiting\" is invalid"
    get:
      summary: returns the list status entries associated with the specimen with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: specimen_key
        in: path
        description: key of the Specimen
        required: true
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/status_entry'
        403:
          description: Unauthorized (can be because of the Specimen is marked as inactive, aka discarded, aka deleted)
        404:
          description: Not Found
  /status/key/{status_key}:
    get:
      summary: returns the status entry with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: status_key
        in: path
        description: key of the Status Entry
        required: true
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/status_entry'
        403:
          description: Unauthorized (can be because of the associated Specimen is marked as inactive, aka discarded, aka deleted)
        404:
          description: Not Found
    delete:
      summary: deletes the status entry with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: status_key
        in: path
        description: key of the Status Entry
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        200:
          description: Deleted
        403:
          description: Unauthorized (can be because of the associated Specimen is marked as inactive, aka discarded, aka deleted)
        404:
          description: Not Found
    put:
      summary: updates the status entry with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: status_key
        in: path
        description: key of the Status Entry
        required: true
        type: integer
      - in: body
        name: Status Entry
        required: true
        description: Status Entry
        schema:
          $ref: '#/definitions/status_entry'
      produces:
      - text/plain
      responses:
        200:
          description: Updated
        403:
          description: Unauthorized (can be because of the associated Specimen is marked as inactive, aka discarded, aka deleted)
        404:
          description: Not Found
        409:
          description: "Conflict, descriptive error message detailing if the stage_name is invalid."
          schema:
            example: "stage_name \"Fruiting\" is invalid"
definitions:
  user:
    type: object
    properties:
      user_key:
        description: "Nullable.\nReadonly value. (automatically generated on POST method)"
        type: integer
        example: 0
      username:
        description: ""
        type: string
        example: Bob
      password:
        description: "Nullable (used only in POST method)."
        type: string
        example: kolbiharcos1234
      user_type:
        description: "Nullable.\nReadonly value.\nAdmin or User (or possibly other future roles)"
        type: string
        example: User
...
  status_entry:
    type: object
    properties:
      entry_key:
        description: "Nullable.\nReadonly value. (automatically generated on POST method)"
        type: integer
        example: 0
      entry_time:
        description: "Nullable (default value is NOW in POST method).\nUnix datetime format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
        type: integer
        example: 1492992000000
      stage_name:
        description: "Stage of the developement (must be from the predefined list)"
        type: string
        example: "Fruiting"
      specimen_key:
        description: "Readonly value (can only be set in POST method).\nKey of the associated specimen."
        type: integer
        example: 145